Cleaned up AutoIncrementingMBean / AutoIncrementing
